+++ beacon/Cargo.toml
@@ -14,8 +14,8 @@ plain_hasher = { version = "0.2", default-features = false }
 impl-serde = { version = "0.1", optional = true }
 serde = { version = "1.0", optional = true }
 serde_derive = { version = "1.0", optional = true }
-ssz = { version = "0.1.1", path = "../util/ssz", default-features = false }
-ssz-derive = { version = "0.1.1", path = "../util/ssz-derive" }
+ssz = { version = "0.1.1", path = "../utils/ssz", default-features = false }
+ssz-derive = { version = "0.1.1", path = "../utils/ssz-derive" }
 fixed-hash = { version = "0.3.0", default-features = false }
 parity-codec = { version = "3.2", default-features = false, optional = true, features = ["derive"] }
 digest = "0.8"
