+++ beacon/src/utils.rs
