+++ utils/shuffling/src/lib.rs
