+++ utils/ssz-derive/src/encode.rs
