+++ utils/ssz-derive/src/hash.rs
