+++ yamltests/Cargo.toml
@@ -9,6 +9,6 @@ serde = "1.0"
 serde_derive = "1.0"
 serde_yaml = "0.8"
 clap = "2.32"
-ssz = { path = "../util/ssz" }
+ssz = { path = "../utils/ssz" }
 primitives = { package = "shasper-primitives", path = "../primitives" }
 beacon = { path = "../beacon" }
